Changed Player input
I did as much as I could without assets and being able to compile, a lot
of the controllers programming is going to be trial and error. We see if
the way the player moves is desired and if not we tweak it. The same
will go for the missiles.